django-3/web/requirements.txt
pillow
django-crum